feat(server): 添加S3存储功能并重构用户认证流程
- 初始化S3客户端用于文件存储
- 在启动时配置S3存储服务
- 重构OAuth回调URL配置，修正Google和GitHub回调地址
- 配置AWS S3访问凭证和存储桶参数

refactor(controller): 统一用户ID获取和会话管理逻辑

- 创建工具类utils来统一处理用户ID和会话信息
- 替换原有的Sa-Token用户认证获取方式
- 优化用户控制器中的身份验证流程
- 改进文件上传控制器的用户ID处理逻辑

feat(user): 添加用户个人资料更新功能

- 实现UpdateProfile方法允许用户修改个人信息
- 支持更新用户名、头像、手机号、性别、生日等字段
- 集成表单验证和错误处理机制
- 自动刷新缓存确保数据一致性

refactor(model): 调整用户模型ID类型为int64

- 将SysUser模型的ID字段从uint改为int64
- 更新GetUserByID方法参数类型
- 适配相关缓存和查询逻辑

fix(s3): 移除对象存储中的ACL配置

- 删除S3上传中的ACL参数设置
- 改用存储桶策略控制访问权限
- 简化上传接口参数

+++ pkg/server/model/user.go
@@ -11,7 +11,7 @@ import (
 )
 
 type SysUser struct {
-	ID         uint       `json:"id" gorm:"primarykey;column:id"`
+	ID         int64      `json:"id" gorm:"primarykey;column:id"`
 	CreateTime time.Time  `json:"create_time" gorm:"column:create_time;autoCreateTime"`
 	UpdateTime time.Time  `json:"update_time" gorm:"column:update_time;autoUpdateTime"`
 	Username   string     `json:"username" gorm:"column:username;not null"`
@@ -41,7 +41,7 @@ const (
 
 // GetUserByID 根据用户ID获取用户信息（带Redis缓存）
 // 缓存策略：先查Redis，未命中则查数据库并写入缓存
-func GetUserByID(db *gorm.DB, userID uint) (*SysUser, error) {
+func GetUserByID(db *gorm.DB, userID int64) (*SysUser, error) {
 	// 1. 尝试从Redis缓存获取
 	cacheKey := GetUserCacheKey(userID)
 	cachedData, err := redis.Get(cacheKey)
@@ -75,13 +75,13 @@ func GetUserByID(db *gorm.DB, userID uint) (*SysUser, error) {
 }
 
 // InvalidateUserCache 使指定用户的缓存失效
-func InvalidateUserCache(userID uint) error {
+func InvalidateUserCache(userID int64) error {
 	cacheKey := GetUserCacheKey(userID)
 	return redis.Del(cacheKey)
 }
 
 // GetUserCacheKey 生成用户缓存键
-func GetUserCacheKey(userID uint) string {
+func GetUserCacheKey(userID int64) string {
 	return fmt.Sprintf("%s%d", UserCachePrefix, userID)
 }
